Update to swift 3
