+++ README.md
@@ -1,4 +1,5 @@
 # SwiftPokerHands
+![](https://img.shields.io/badge/swift-3.0-orange.svg)
 A simple five card poker hand classifier
 
 ### Examples:
@@ -7,11 +8,11 @@ A simple five card poker hand classifier
 
 ```
 let cards = [
-    PokerCard(rank: .Four, suit: .Hearts),
-    PokerCard(rank: .Five, suit: .Diamonds),
-    PokerCard(rank: .Six, suit: .Clubs),
-    PokerCard(rank: .Seven, suit: .Spades),
-    PokerCard(rank: .Eight, suit: .Hearts),
+    Card(rank: .four, suit: .hearts),
+    Card(rank: .five, suit: .diamonds),
+    Card(rank: .six, suit: .clubs),
+    Card(rank: .seven, suit: .spades),
+    Card(rank: .eight, suit: .hearts),
 ]
 let hand = PokerHand(cards: cards)
 print(hand.handRank()) // Straight
@@ -21,18 +22,18 @@ print(hand.handRank()) // Straight
 
 ```
 let cards = [
-    PokerCard(rank: .Four, suit: .Hearts),
-    PokerCard(rank: .Four, suit: .Diamonds),
-    PokerCard(rank: .Five, suit: .Hearts),
-    PokerCard(rank: .Five, suit: .Hearts),
-    PokerCard(rank: .Six, suit: .Hearts),
+    Card(rank: .four, suit: .hearts),
+    Card(rank: .four, suit: .diamonds),
+    Card(rank: .five, suit: .hearts),
+    Card(rank: .five, suit: .hearts),
+    Card(rank: .six, suit: .hearts),
 ]
 let cards2 = [
-    PokerCard(rank: .Five, suit: .Hearts),
-    PokerCard(rank: .Five, suit: .Hearts),
-    PokerCard(rank: .Seven, suit: .Hearts),
-    PokerCard(rank: .Six, suit: .Hearts),
-    PokerCard(rank: .Six, suit: .Clubs),
+    Card(rank: .five, suit: .hearts),
+    Card(rank: .five, suit: .hearts),
+    Card(rank: .seven, suit: .hearts),
+    Card(rank: .six, suit: .hearts),
+    Card(rank: .six, suit: .clubs),
 ]
 let hand = PokerHand(cards: cards)
 let hand2 = PokerHand(cards: cards2)
@@ -44,25 +45,25 @@ print(hand < hand2) // true
 
 ```
 let cards = [
-    PokerCard(rank: .Four, suit: .Hearts),
-    PokerCard(rank: .Four, suit: .Diamonds),
-    PokerCard(rank: .Five, suit: .Hearts),
-    PokerCard(rank: .Five, suit: .Hearts),
-    PokerCard(rank: .Six, suit: .Hearts),
+    Card(rank: .four, suit: .hearts),
+    Card(rank: .four, suit: .diamonds),
+    Card(rank: .five, suit: .hearts),
+    Card(rank: .five, suit: .hearts),
+    Card(rank: .six, suit: .hearts),
 ]
 let cards2 = [
-    PokerCard(rank: .Five, suit: .Hearts),
-    PokerCard(rank: .Five, suit: .Hearts),
-    PokerCard(rank: .Seven, suit: .Hearts),
-    PokerCard(rank: .Six, suit: .Hearts),
-    PokerCard(rank: .Six, suit: .Clubs),
+    Card(rank: .five, suit: .hearts),
+    Card(rank: .five, suit: .hearts),
+    Card(rank: .seven, suit: .hearts),
+    Card(rank: .six, suit: .hearts),
+    Card(rank: .six, suit: .clubs),
 ]
 let cards3 = [
-    PokerCard(rank: .Four, suit: .Hearts),
-    PokerCard(rank: .Five, suit: .Diamonds),
-    PokerCard(rank: .Six, suit: .Clubs),
-    PokerCard(rank: .Seven, suit: .Spades),
-    PokerCard(rank: .Eight, suit: .Hearts),
+    Card(rank: .four, suit: .hearts),
+    Card(rank: .five, suit: .diamonds),
+    Card(rank: .six, suit: .clubs),
+    Card(rank: .seven, suit: .spades),
+    Card(rank: .eight, suit: .hearts),
 ]
 let hand = PokerHand(cards: cards)
 let hand2 = PokerHand(cards: cards2)
@@ -71,7 +72,7 @@ let hand3 = PokerHand(cards: cards3)
 print(hand < hand2) // true
 print(hand2.handRank()) // TwoPair
 
-print(hand.suits) // [Suit.Diamonds: 1, Suit.Hearts: 4]
+print(hand.suits) // [Suit.diamonds: 1, Suit.hearts: 4]
 print(hand.handRank()) // TwoPair
 print(hand3.handRank()) // Straight
 ```
