+++ SwiftPokerHands.playground/contents.xcplayground
@@ -1,2 +1,4 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<playground version='6.0' target-platform='osx' requires-full-environment='true'/>
+<playground version='5.0' target-platform='osx' last-migration='0800'>
+    <timeline fileName='timeline.xctimeline'/>
+</playground>
